+++ x/gov/keeper/msg_server.go
@@ -97,8 +97,8 @@ func (k msgServer) SubmitProposal(ctx context.Context, msg *v1.MsgSubmitProposal
 	}
 
 	// ref: https://github.com/cosmos/cosmos-sdk/issues/9683
-	k.environment.GasService.GetGasMeter(ctx).Consume(
-		3*k.environment.GasService.GetGasConfig(ctx).WriteCostPerByte*uint64(len(bytes)),
+	k.GasService.GasMeter(ctx).Consume(
+		3*k.GasService.GasConfig(ctx).WriteCostPerByte*uint64(len(bytes)),
 		"submit proposal",
 	)
 
@@ -108,7 +108,7 @@ func (k msgServer) SubmitProposal(ctx context.Context, msg *v1.MsgSubmitProposal
 	}
 
 	if votingStarted {
-		if err := k.environment.EventService.EventManager(ctx).EmitKV(
+		if err := k.EventService.EventManager(ctx).EmitKV(
 			govtypes.EventTypeSubmitProposal,
 			event.NewAttribute(govtypes.AttributeKeyVotingPeriodStart, fmt.Sprintf("%d", proposal.Id)),
 		); err != nil {
@@ -171,7 +171,7 @@ func (k msgServer) CancelProposal(ctx context.Context, msg *v1.MsgCancelProposal
 		return nil, err
 	}
 
-	if err := k.environment.EventService.EventManager(ctx).EmitKV(
+	if err := k.EventService.EventManager(ctx).EmitKV(
 		govtypes.EventTypeCancelProposal,
 		event.NewAttribute(sdk.AttributeKeySender, msg.Proposer),
 		event.NewAttribute(govtypes.AttributeKeyProposalID, fmt.Sprint(msg.ProposalId)),
@@ -181,8 +181,8 @@ func (k msgServer) CancelProposal(ctx context.Context, msg *v1.MsgCancelProposal
 
 	return &v1.MsgCancelProposalResponse{
 		ProposalId:     msg.ProposalId,
-		CanceledTime:   k.environment.HeaderService.GetHeaderInfo(ctx).Time,
-		CanceledHeight: uint64(k.environment.HeaderService.GetHeaderInfo(ctx).Height),
+		CanceledTime:   k.HeaderService.HeaderInfo(ctx).Time,
+		CanceledHeight: uint64(k.HeaderService.HeaderInfo(ctx).Height),
 	}, nil
 }
 
@@ -293,7 +293,7 @@ func (k msgServer) Deposit(ctx context.Context, msg *v1.MsgDeposit) (*v1.MsgDepo
 	}
 
 	if votingStarted {
-		if err := k.environment.EventService.EventManager(ctx).EmitKV(
+		if err := k.EventService.EventManager(ctx).EmitKV(
 			govtypes.EventTypeProposalDeposit,
 			event.NewAttribute(govtypes.AttributeKeyVotingPeriodStart, fmt.Sprintf("%d", msg.ProposalId)),
 		); err != nil {
@@ -374,13 +374,13 @@ func (k msgServer) SudoExec(ctx context.Context, msg *v1.MsgSudoExec) (*v1.MsgSu
 	}
 
 	var msgResp protoiface.MessageV1
-	if err := k.environment.BranchService.Execute(ctx, func(ctx context.Context) error {
+	if err := k.BranchService.Execute(ctx, func(ctx context.Context) error {
 		// TODO add route check here
-		if err := k.environment.RouterService.MessageRouterService().CanInvoke(ctx, sdk.MsgTypeURL(sudoedMsg)); err != nil {
+		if err := k.RouterService.MessageRouterService().CanInvoke(ctx, sdk.MsgTypeURL(sudoedMsg)); err != nil {
 			return errors.Wrapf(govtypes.ErrInvalidProposal, err.Error())
 		}
 
-		msgResp, err = k.environment.RouterService.MessageRouterService().InvokeUntyped(ctx, sudoedMsg)
+		msgResp, err = k.RouterService.MessageRouterService().InvokeUntyped(ctx, sudoedMsg)
 		if err != nil {
 			return errors.Wrapf(err, "failed to execute sudo-ed message; message %v", sudoedMsg)
 		}