+++ x/group/keeper/proposal_executor.go
@@ -17,7 +17,7 @@ import (
 // doExecuteMsgs routes the messages to the registered handlers. Messages are limited to those that require no authZ or
 // by the account of group policy only. Otherwise this gives access to other peoples accounts as the sdk middlewares are bypassed
 func (k Keeper) doExecuteMsgs(ctx context.Context, proposal group.Proposal, groupPolicyAcc sdk.AccAddress, decisionPolicy group.DecisionPolicy) error {
-	currentTime := k.environment.HeaderService.GetHeaderInfo(ctx).Time
+	currentTime := k.HeaderService.HeaderInfo(ctx).Time
 
 	// Ensure it's not too early to execute the messages.
 	minExecutionDate := proposal.SubmitTime.Add(decisionPolicy.GetMinExecutionPeriod())
@@ -45,7 +45,7 @@ func (k Keeper) doExecuteMsgs(ctx context.Context, proposal group.Proposal, grou
 	}
 
 	for i, msg := range msgs {
-		if _, err := k.environment.RouterService.MessageRouterService().InvokeUntyped(ctx, msg); err != nil {
+		if _, err := k.RouterService.MessageRouterService().InvokeUntyped(ctx, msg); err != nil {
 			return errorsmod.Wrapf(err, "message %s at position %d", sdk.MsgTypeURL(msg), i)
 		}
 	}